requirements.txt
google-generativeai
edge-tts
moviepy
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
